fix: 필터 UX 개선 및 중복 코드 정리 #178
